requirements.txt
appdirs==1.4.3
args==0.1.0
clint==0.5.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
xlwt==1.2.0
